add streaming to integration test
